cligen-#116 Add `write_version` target to Makefile
Introduced a new `write_version` target to the Makefile, which generates a version file based on the latest Git tag. The `version` file is added to both the `all` and `clean` targets to ensure it is created during build and removed during cleanup.

cligen-#116 Set `fetch-depth: 0` for checkout steps in CI workflow

Ensure the full git history is fetched in the CI workflow by setting `fetch-depth: 0` for all `actions/checkout@v4` steps. This change is necessary for operations that require complete commit history.