Remove invalid structure
C spec 6.2.5-20 says that a structure cannot be zero size:

A structure type describes a sequentially allocated nonempty set of member objects
(and, in certain circumstances, an incomplete array), each of which has an optionally
specified name and possibly distinct type.

Commenting it out doesn't seem to break anything build-wise.